change style of select teg, add populate script for db. change cascade Type for some entity

+++ src/main/java/javaonline/web/admin/OrderController.java
@@ -1,16 +1,15 @@
 package javaonline.web.admin;
 
+import javaonline.dao.ICookedDishDao;
 import javaonline.dao.IDishDao;
 import javaonline.dao.IEmployeeDao;
 import javaonline.dao.IOrderDao;
-import javaonline.dao.entity.Dish;
-import javaonline.dao.entity.Employee;
-import javaonline.dao.entity.Order;
-import javaonline.dao.entity.OrderStatus;
+import javaonline.dao.entity.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.ServletRequestDataBinder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
@@ -19,22 +18,29 @@ import javax.servlet.http.HttpServletRequest;
 import java.beans.PropertyEditorSupport;
 import java.sql.Date;
 import java.text.SimpleDateFormat;
+import java.util.List;
 
 @Controller
 public class OrderController {
 
     private IOrderDao ordersDaoService;
-    private IEmployeeDao IEmployeeDao;
-    private IDishDao IDishDao;
+    private IEmployeeDao employeeDao;
+    private IDishDao dishDao;
+    private ICookedDishDao cookedDishDao;
 
     @Autowired
-    public void setIDishDao(IDishDao IDishDao) {
-        this.IDishDao = IDishDao;
+    public void setCookedDishDao(ICookedDishDao cookedDishDao) {
+        this.cookedDishDao = cookedDishDao;
     }
 
     @Autowired
-    public void setIEmployeeDao(IEmployeeDao IEmployeeDao) {
-        this.IEmployeeDao = IEmployeeDao;
+    public void setDishDao(IDishDao dishDao) {
+        this.dishDao = dishDao;
+    }
+
+    @Autowired
+    public void setEmployeeDao(IEmployeeDao employeeDao) {
+        this.employeeDao = employeeDao;
     }
 
     @Autowired
@@ -47,13 +53,13 @@ public class OrderController {
         binder.registerCustomEditor(Employee.class, "employeeId", new PropertyEditorSupport() {
             @Override
             public void setAsText(String text) {
-                setValue(IEmployeeDao.getEmployeeById(Integer.valueOf(text)));
+                setValue(employeeDao.getEmployeeById(Integer.valueOf(text)));
             }
         });
         binder.registerCustomEditor(Dish.class, "dishes", new PropertyEditorSupport() {
             @Override
             public void setAsText(String text) {
-                setValue(IDishDao.getDishByName(text));
+                setValue(dishDao.getDishByName(text));
             }
         });
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
@@ -78,13 +84,43 @@ public class OrderController {
     @RequestMapping(value = "/admin/addOrder", method = RequestMethod.GET)
     public String addOrder(Model model) {
         model.addAttribute("order", new Order());
-        model.addAttribute("listOfEmployee", IEmployeeDao.getAllEmployees());
-        model.addAttribute("listOfDishes", IDishDao.getAllDishes());
+        model.addAttribute("listOfEmployee", employeeDao.getWaiters());
+        model.addAttribute("listOfDishes", dishDao.getAllDishes());
         return "admin/order/addOrUpdateOrder";
     }
 
     @RequestMapping(value = "/admin/addOrUpdateOrder", method = RequestMethod.POST)
-    public String addNewDish(@ModelAttribute("Order") Order order, final RedirectAttributes redirectAttributes) {
+    public String addNewDish(@ModelAttribute Order order, BindingResult result,
+                             Model model, final RedirectAttributes redirectAttributes) {
+
+        boolean error = false;
+
+        if (order.getDishes() == null) {
+            result.rejectValue("dishes", "error.Dishes");
+            error = true;
+        }
+
+        if (order.getEmployeeId() == null) {
+            result.rejectValue("employeeId", "error.EmployeeId");
+            error = true;
+        }
+
+        if (order.getTableNumber() <= 0){
+            result.rejectValue("tableNumber", "error.TableNumber");
+            error = true;
+        }
+
+        if (order.getDate() == null) {
+            result.rejectValue("date", "error.Date");
+            error = true;
+        }
+
+        if (error) {
+            model.addAttribute("listOfEmployee", employeeDao.getWaiters());
+            model.addAttribute("listOfDishes", dishDao.getAllDishes());
+            return "admin/order/addOrUpdateOrder";
+        }
+
         redirectAttributes.addFlashAttribute("css", "success");
         if (ordersDaoService.getOrderById(order.getId())==null){
             redirectAttributes.addFlashAttribute("msg", "Order was added!");
@@ -100,8 +136,8 @@ public class OrderController {
     public String updateOrder(@PathVariable int id,Model model) {
         Order order = ordersDaoService.getOrderById(id);
         model.addAttribute("order", order);
-        model.addAttribute("listOfEmployee", IEmployeeDao.getAllEmployees());
-        model.addAttribute("listOfDishes", IDishDao.getAllDishes());
+        model.addAttribute("listOfEmployee", employeeDao.getAllEmployees());
+        model.addAttribute("listOfDishes", dishDao.getAllDishes());
         return "admin/order/addOrUpdateOrder";
     }
 
@@ -118,10 +154,16 @@ public class OrderController {
     @RequestMapping(value = "/admin/order/{id}/delete", method = RequestMethod.GET)
     public String delete(@PathVariable int id, final RedirectAttributes redirectAttributes) {
         Order order = ordersDaoService.getOrderById(id);
-        order.getDishes().clear();
-        redirectAttributes.addFlashAttribute("css", "success");
-        redirectAttributes.addFlashAttribute("msg", "Order was deleted!");
-        ordersDaoService.deleteOrder(order);
+        List<CookedDish> cookedDishes = cookedDishDao.getCookedDishesByOrderId(order.getId());
+        if (cookedDishes==null){
+            order.getDishes().clear();
+            ordersDaoService.deleteOrder(order);
+            redirectAttributes.addFlashAttribute("css", "success");
+            redirectAttributes.addFlashAttribute("msg", "Order was deleted!");
+        }else{
+            redirectAttributes.addFlashAttribute("css", "danger");
+            redirectAttributes.addFlashAttribute("msg", "Order can't be deleted! Some dishes has already cooked");
+        }
         return "redirect:/admin/getAllOrders";
     }
